Fixes from an attempted winelib build.
The winelib headers don't have GWL_foo, only GWLP_foo (which, fair
enough, I should have been using already). And a side effect was to
point out some slightly incautious integer types in printf argument
lists.

+++ windows/winser.c
@@ -125,10 +125,12 @@ static char *serial_configure(Serial *serial, HANDLE serport, Conf *conf)
          * Configurable parameters.
          */
         dcb.BaudRate = conf_get_int(conf, CONF_serspeed);
-        logeventf(serial->logctx, "Configuring baud rate %lu", dcb.BaudRate);
+        logeventf(serial->logctx, "Configuring baud rate %lu",
+                  (unsigned long)dcb.BaudRate);
 
         dcb.ByteSize = conf_get_int(conf, CONF_serdatabits);
-        logeventf(serial->logctx, "Configuring %u data bits", dcb.ByteSize);
+        logeventf(serial->logctx, "Configuring %u data bits",
+                  (unsigned)dcb.ByteSize);
 
         switch (conf_get_int(conf, CONF_serstopbits)) {
           case 2: dcb.StopBits = ONESTOPBIT; str = "1 stop bit"; break;
